refactor(web): decompose the Layout shell
After #768 moved SystemStatus and LanguageSelector out, Layout was still
a 991-line component. It is now the composition only (98 lines, under
the max-lines-per-function cap, so its eslint-disable goes); the rest
lives under components/layout/:

- useDiscrepancyModal.ts: the polled discovery status behind the Models
  badge, the modal's open state, the serialised retest and the
  retest-all walk with its cancel, the confirm-before-claim dismiss /
  dismiss-all / dismiss-everything / unpin handlers, the journal ack and
  the SSE badge refresh; returns the modal's props
- useNavigation.ts: the nav items for the current identity, the
  sub-mode map, the sub-mode toggle click, isActive
- useCircuitBreakerStatus.ts: the polled status and its SSE
  invalidations; useLogout.ts: the one logout path, also used by idle
  auto-logout
- SidebarNav.tsx, SidebarFooter.tsx, FailoverNavBadge.tsx,
  DiscoveryNavBadge.tsx: the item list with its two badges, and the
  footer with the logout confirmation

No behaviour change: same data-testids, queries, event listeners and
handler semantics, every comment travelling with its code;
useDiscrepancies is still called unconditionally, which its session
counter depends on. The 6 Layout suites pass unchanged; DiscoveryNavBadge
gets its own test (count vs dot named by the label; opens on Enter,
Space and click). The file leaves the size allowlist (1065 -> 98).

Verified on the rebuilt dev stack: the 11 nav items render, a second
click on Chat flips it to Conversation, the discovery badge opens the
modal from the keyboard and hides while it is open, the theme toggles,
the logout button asks for confirmation, no console errors.
